# car-management-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/cars": {
      "get": {
        "tags": [
          "car-controller"
        ],
        "operationId": "getAllCars",
        "parameters": [
          {
            "name": "carMake",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "garageId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "fromYear",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "toYear",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resources found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseCarDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "post": {
        "tags": [
          "car-controller"
        ],
        "operationId": "createCar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCarDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "200": {
            "description": "Resource created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCarDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseGarageDTO": {
        "required": [
          "capacity",
          "city",
          "id",
          "location",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ResponseCarDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "productionYear": {
            "type": "integer",
            "format": "int32"
          },
          "licensePlate": {
            "type": "string"
          },
          "garages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseGarageDTO"
            }
          }
        }
      },
...
      "CreateCarDTO": {
        "type": "object",
        "properties": {
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "productionYear": {
            "type": "integer",
            "format": "int32"
          },
          "licensePlate": {
            "type": "string"
          },
          "garageIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
    }
  }
}